move `run_dns_and_pkarr_servers` to `iroh::net::test_utils`
to properly test that we have fixed the discovery bug, we need to run the test dns and pkarr relay servers, so they should be moved to `test_utils` so the code can be shared
